Refactor modal from related items to separate component
