+++ src/features/RelatedItemsAndComparison/ComparisonModal/ComparisonModal.module.css
@@ -1,0 +1,15 @@
+.modal {
+  position: fixed;
+  top: 30vh;
+  left: 10%;
+  width: 80%;
+  z-index: 100;
+  overflow: hidden;
+}
+
+@media (min-width: 768px) {
+  .modal {
+    left: calc(50% - 20rem);
+    width: 40rem;
+  }
+}
