+++ src/features/RelatedItemsAndComparison/RelatedProductsList/RelatedItemCard/RelatedItemCard.module.css
@@ -45,19 +45,3 @@
   z-index: 10;
   background: rgba(0, 0, 0, 0.75);
 }
-
-.modal {
-  position: fixed;
-  top: 30vh;
-  left: 10%;
-  width: 80%;
-  z-index: 100;
-  overflow: hidden;
-}
-
-@media (min-width: 768px) {
-  .modal {
-    left: calc(50% - 20rem);
-    width: 40rem;
-  }
-}
